move batch stuff under the batch package
