+++ src/main/java/uk/gov/hmcts/reform/ref/pup/batch/exception/AppConfigurationException.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.ref.pup.exception;
+package uk.gov.hmcts.reform.ref.pup.batch.exception;
 
 public class AppConfigurationException extends RuntimeException {
