+++ src/main/java/uk/gov/hmcts/reform/ref/pup/batch/exception/FileStorageException.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.ref.pup.exception;
+package uk.gov.hmcts.reform.ref.pup.batch.exception;
 
 public class FileStorageException extends RuntimeException {
     public FileStorageException(Exception e) {
