+++ src/main/java/uk/gov/hmcts/reform/ref/pup/batch/services/azure/AzureFileStorageClient.java
@@ -1,7 +1,7 @@
-package uk.gov.hmcts.reform.ref.pup.services.azure;
+package uk.gov.hmcts.reform.ref.pup.batch.services.azure;
 
-import uk.gov.hmcts.reform.ref.pup.exception.FileStorageException;
-import uk.gov.hmcts.reform.ref.pup.services.file.FileStorageClient;
+import uk.gov.hmcts.reform.ref.pup.batch.exception.FileStorageException;
+import uk.gov.hmcts.reform.ref.pup.batch.services.file.FileStorageClient;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
@@ -9,8 +9,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
 import com.microsoft.azure.storage.StorageException;
-import com.microsoft.azure.storage.blob.CloudBlobContainer;
-import com.microsoft.azure.storage.blob.CloudBlockBlob;
+import com.microsoft.azure.storage.file.CloudFile;
+import com.microsoft.azure.storage.file.CloudFileShare;
 
 import java.io.IOException;
 import java.io.OutputStream;
@@ -19,22 +19,23 @@ import java.util.UUID;
 
 import javax.annotation.PostConstruct;
 
-@Service(value = "azureBlobStorageClient")
+@Service(value = "azureFileStorageClient")
 @ConditionalOnProperty("azure.storage.connection-string")
-public class AzureBlobStorageClient implements FileStorageClient {
+public class AzureFileStorageClient implements FileStorageClient {
 
     @Autowired
-    private CloudBlobContainer cloudBlobContainer;
+    private CloudFileShare cloudFileShare;
 
     @PostConstruct
     void init() throws StorageException {
-        cloudBlobContainer.createIfNotExists();
+        cloudFileShare.createIfNotExists();
     }
 
     @Override
     public void uploadFile(UUID uuid, MultipartFile multipartFile) {
         try {
-            getCloudFile(uuid).upload(multipartFile.getInputStream(), multipartFile.getSize());
+            CloudFile cloudFile = getCloudFile(uuid);
+            cloudFile.upload(multipartFile.getInputStream(), multipartFile.getSize());
         } catch (URISyntaxException | StorageException | IOException e) {
             throw new FileStorageException(e);
         }
@@ -43,7 +44,8 @@ public class AzureBlobStorageClient implements FileStorageClient {
     @Override
     public void streamFileContent(UUID uuid, OutputStream outputStream) {
         try {
-            getCloudFile(uuid).download(outputStream);
+            CloudFile cloudFile = getCloudFile(uuid);
+            cloudFile.download(outputStream);
         } catch (URISyntaxException | StorageException e) {
             throw new FileStorageException(e);
         }
@@ -52,13 +54,14 @@ public class AzureBlobStorageClient implements FileStorageClient {
     @Override
     public void deleteFile(UUID uuid) {
         try {
-            getCloudFile(uuid).delete();
+            CloudFile cloudFile = getCloudFile(uuid);
+            cloudFile.delete();
         } catch (URISyntaxException | StorageException e) {
             throw new FileStorageException(e);
         }
     }
 
-    private CloudBlockBlob getCloudFile(UUID uuid) throws StorageException, URISyntaxException {
-        return cloudBlobContainer.getBlockBlobReference(uuid.toString());
+    private CloudFile getCloudFile(UUID uuid) throws StorageException, URISyntaxException {
+        return cloudFileShare.getRootDirectoryReference().getFileReference(uuid.toString());
     }
 }
