+++ src/main/java/uk/gov/hmcts/reform/ref/pup/controller/PackagedBankAccountController.java
@@ -2,7 +2,7 @@ package uk.gov.hmcts.reform.ref.pup.controller;
 
 import uk.gov.hmcts.reform.ref.pup.domain.PaymentAccount;
 import uk.gov.hmcts.reform.ref.pup.domain.ProfessionalUser;
-import uk.gov.hmcts.reform.ref.pup.dto.PaymentAccountDto;
+import uk.gov.hmcts.reform.ref.pup.dto.PaymentAccountRequest;
 import uk.gov.hmcts.reform.ref.pup.exception.ApplicationException;
 import uk.gov.hmcts.reform.ref.pup.services.PaymentAccountService;
 
@@ -41,7 +41,7 @@ public class PackagedBankAccountController {
     @ApiResponses(value = { 
             @ApiResponse(code = 200, message = "Success", response = ProfessionalUser.class) 
     })
-    public ResponseEntity<PaymentAccount> createPaymentAccount(@RequestBody @Valid PaymentAccountDto paymentAccount) throws ApplicationException {
+    public ResponseEntity<PaymentAccount> createPaymentAccount(@RequestBody @Valid PaymentAccountRequest paymentAccount) throws ApplicationException {
         return ResponseEntity.ok(paymentAccountService.create(paymentAccount.getPbaNumber(), paymentAccount.getOrganisationId()));
     }
