+++ src/main/java/uk/gov/hmcts/reform/ref/pup/exception/ApplicationException.java
@@ -2,6 +2,7 @@ package uk.gov.hmcts.reform.ref.pup.exception;
 
 public class ApplicationException extends Exception {
 
+    private static final long serialVersionUID = 1L;
 
     public ApplicationException(String message) {
         super(message);
